GetUser method moved to DataPreparationHelper class.  as per PR comment
