+++ Saucedemo/Saucedemo/Tests/CartTest.cs
@@ -13,7 +13,7 @@ namespace Saucedemo.Tests
         [Test]
         public void Test1()
         {
-            var user = GetUser();
+            var user = DataPreparationHelper.GetUser();
 
             var cartPage = new LoginPage(ChromeDriver, true)
                .SuccessfulLogin(user)
